Add Basic Authentication

+++ 07/hajiboot2-tweeter/src/main/java/hajiboot/tweet/web/TweetController.java
@@ -4,6 +4,7 @@ import java.net.URI;
 import java.time.Clock;
 import java.time.Instant;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -20,7 +21,9 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -51,15 +54,27 @@ public class TweetController {
 		}
 	}
 
+	@GetMapping(path = "tweets")
+	public ResponseEntity<List<TweetOutput>> searchTweets(@RequestParam String text) {
+		return ResponseEntity.ok().build();
+	}
+
 	@DeleteMapping(path = "tweets/{uuid}")
-	public ResponseEntity<Void> deleteTweet(@PathVariable UUID uuid) {
-		this.tweetMapper.deleteByUuid(uuid);
-		return ResponseEntity.noContent().build();
+	public ResponseEntity<?> deleteTweet(@PathVariable UUID uuid, @RequestAttribute("tweeter") Tweeter tweeter) {
+		final Tweet tweet = this.tweetMapper.findByUuid(uuid);
+		if (tweet.isOwnedBy(tweeter)) {
+			this.tweetMapper.deleteByUuid(uuid);
+			return ResponseEntity.noContent().build();
+		}
+		else {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN)
+					.body(Map.of("error", "Forbidden", "status", 403));
+		}
 	}
 
 	@PostMapping(path = "tweets")
-	public ResponseEntity<TweetOutput> postTweets(@Validated @RequestBody TweetInput input, UriComponentsBuilder builder) {
-		final Tweet tweet = new Tweet(this.idGenerator.generateId(), input.getText(), new Tweeter(input.getUsername()), Instant.now(this.clock));
+	public ResponseEntity<TweetOutput> postTweets(@Validated @RequestBody TweetInput input, @RequestAttribute("tweeter") Tweeter tweeter, UriComponentsBuilder builder) {
+		final Tweet tweet = new Tweet(this.idGenerator.generateId(), input.getText(), new Tweeter(tweeter.getUsername()), Instant.now(this.clock));
 		this.tweetMapper.insert(tweet);
 		final TweetOutput output = TweetOutput.fromTweet(tweet);
 		final URI location = builder.path("tweets/{uuid}").build(output.getUuid());
